% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{adjust_load}
\alias{adjust_load}
\title{Load to tension}
\usage{
adjust_load(load, angle)
}
\arguments{
\item{load}{Numeric vector of load value or values to adjust.}

\item{angle}{The included angle across the block in degrees.}
}
\value{
A numeric vector representing the adjusted load value.
}
\description{
Calculates the tension on a line travelling through a block based on the load
on the block and included angle of the line over the block.
}
\author{
Bill DeVoe, \email{William.DeVoe@maine.gov}
}
