% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{export_loadcell}
\alias{export_loadcell}
\title{Export load cell data}
\usage{
export_loadcell(data, prefix)
}
\arguments{
\item{data}{An object of class `LoadCellHauls` or `LoadCellPeaks`}

\item{prefix}{*Optional* A prefix to append to the filenames of the
output CSVs.}
}
\value{
Boolean True if successful.
}
\description{
For an input object of class `LoadCellHauls` or `LoadCellPeaks`, flattens the
 object data into CSVs of data. Optionally, a prefix can be given to be added
 to the filename. The following CSVs are created:
\describe{
\item{prefix-lc_data.csv}{The load cell data, with UID and haul numbers
added.}
\item{prefix-lc_haul.csv}{The haul attributes, including the settings used
for peak analysis. UID is included to relate hauls to load cell data.}
\item{prefix-lc_peaks.csv}{The actual peak values for each haul, only if peak
 analysis was successful for the haul. Includes UID and haul number.}
}
}
\author{
Bill DeVoe, \email{William.DeVoe@maine.gov}
}
