% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{load_csvs}
\alias{load_csvs}
\title{Load CSVs}
\usage{
load_csvs(dir = getwd())
}
\arguments{
\item{dir}{Directory to scan for load cell CSVs. If no directory is provided,
the current working directory will be used.}
}
\value{
A class of type `LoadCellData`; class items are named with the CSV
filename. Each class item represents a nested list containing the following
attributes:
\describe{
\item{data}{A dataframe of the load cell data contained in the CSV.}
\item{sn}{Serial number of the load cell USB adapter.}
\item{traps}{The number of traps per trawl, if entered in the interface.}
\item{start_dt}{A POSIXlt object representing the start datetime of the CSV.}
\item{end_dt}{A POSIXlt object representing the end datetime of the CSV.}
\item{seconds}{The number of elapsed seconds in the CSV.}
\item{max_load}{An integer representing the maximum load within the CSV.}
\item{min_load}{An integer representing the minimum load within the CSV;
useful for determining a "zero" value for the weight of the block with no
line.}
\item{kfactor}{The Keliher conversion factor applied to the data.}
}
}
\description{
For an input directory, loads CSVs created by the DMR load cell software into
an output class `LoadCellData`.
}
\author{
Bill DeVoe, \email{William.DeVoe@maine.gov}
}
