% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_hauls}
\alias{parse_hauls}
\title{Parse Hauls}
\usage{
parse_hauls(data, split, min_load, min_time, min_gap, pass)
}
\arguments{
\item{data}{An object of class `LoadCellData` representing data loaded from
load cell CSVs.}

\item{split}{Boolean indicating if CSV files should be split into separate
hauls; if a fisherman recorded separate CSVs for each haul, this option would
be False and the hauls will be numbered sequentially per day. If True, hauls
will be parsed from the CSV data according to options set in the remaining
function arguments.}

\item{min_load}{Numeric - If split is True, load cell data below this
threshold will be excluded to create gaps between hauls. If this argument is
omitted, the minimum value for each CSV will be used.}

\item{min_time}{*Optional* Numeric - If split is True, hauls resulting from
excluding load values below `min_load` must be this many seconds in duration
to be considered a haul. This is useful for excluding junk data produced by a
 load cell being jostled in rough seas in between actual hauls.}

\item{min_gap}{*Optional* Numeric - If split is True, a haul must begin this
many seconds after the previous haul to be considered a separate haul.}

\item{pass}{*Optional* Numeric; if the CSV is less than this many seconds in
length, it will not be split and will be assumed to represent a single haul.}
}
\value{
An object of class `LoadCellHauls`. Each item in the class is
represented by a unique name constructed by concaternation of the load cell
serial number, date, and haul number, ie "SN-2018-03-24-1". Each item has
attributes as follows:
\describe{
\item{haul}{The haul number of the data.}
\item{data}{A dataframe containing the load cell data for the haul.}
\item{sn}{Serial number of the load cell adapter.}
\item{traps}{The number of traps per trawl, if entered in the interface.}
\item{start_dt}{A POSIXlt object representing the start datetime of the haul.}
\item{end_dt}{A POSIXlt object representing the end datetime of the haul.}
\item{seconds}{The number of elapsed seconds in the haul.}
\item{max_load}{An integer representing the maximum load within the haul.}
\item{kfactor}{The Keliher conversion factor applied to the data.}
}
}
\description{
For an input object of class `LoadCellData`, parses load cell file data into
hauls by splitting apart load cell data based on minimum load, time
between hauls, and minimum haul length. Return an object of class
`LoadCellHauls`.
}
\author{
Bill DeVoe, \email{William.DeVoe@maine.gov}
}
