% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{adjust_load}
\alias{adjust_load}
\title{Load to tension}
\usage{
adjust_load(data, kfactor)
}
\arguments{
\item{data}{An object of class `LoadCellData`.}

\item{kfactor}{The Keliher conversion factor to apply to the load data.}
}
\value{
An object of class `LoadCellData.`
}
\description{
Adjusts load values from CSV data based on a correction multipler (the
Keliher conversion factor.)
}
\author{
Bill DeVoe, \email{William.DeVoe@maine.gov}
}
